requirements.txt
flask
openai
pinecone-client
python-dotenv
requests
llama-index
gunicorn
flask-cors